node:fs: move fs.watch's FSWatcher into a lazily loaded internal module
fs.watch now delegates to a new internal/fs/watch module that holds the
FSWatcher class, so the class is only evaluated when fs.watch is used. It
gets the native fs binding from internal/fs/binding, the same way the other
lazily loaded internal/fs/* modules do.

The InternalModuleRegistry id -> tag table consumed by the Rust resolver is
code-generated (generated_resolved_source_tag.rs), so adding the module and
rebuilding keeps it in sync; the module is also listed in the
process.binding('natives') table.
