+++ src/js/node/fs.ts
@@ -1,6 +1,5 @@
 // Hardcoded module "node:fs"
 import type { Dirent as DirentType, PathLike, Stats as StatsType } from "fs";
-const EventEmitter = require("node:events");
 const promises = require("node:fs/promises");
 const types = require("node:util/types");
 const {
@@ -41,78 +40,6 @@ function nullcallback(callback) {
   return FunctionPrototypeBind.$call(callback, undefined, null);
 }
 const FunctionPrototypeBind = nullcallback.bind;
-
-class FSWatcher extends EventEmitter {
-  #watcher;
-  #listener;
-  constructor(path, options, listener) {
-    super();
-
-    if (path instanceof URL) {
-      path = Bun.fileURLToPath(path);
-    } else if (typeof path === "string" && path.startsWith("file:")) {
-      path = Bun.fileURLToPath(path);
-    }
-
-    if (typeof options === "function") {
-      listener = options;
-      options = {};
-    } else if (typeof options === "string") {
-      options = { encoding: options };
-    }
-
-    if (typeof listener !== "function") {
-      listener = () => {};
-    }
-
-    this.#listener = listener;
-    try {
-      this.#watcher = fs.watch(path, options || {}, this.#onEvent.bind(this));
-    } catch (e: any) {
-      e.path = path;
-      e.filename = path;
-      throw e;
-    }
-  }
-
-  #onEvent(eventType, filenameOrError) {
-    if (eventType === "close") {
-      // close on next microtask tick to avoid long-running function calls when
-      // we're trying to detach the watcher
-      queueMicrotask(() => {
-        this.emit("close", filenameOrError);
-      });
-      return;
-    } else if (eventType === "error") {
-      // TODO: Next.js/watchpack causes this to emits weird EACCES errors on
-      // paths that shouldn't be watched. A better solution is to figure out why
-      // these paths get watched in the first place. For now we will rewrite the
-      // .code, which will cause their code path to ignore the error.
-      if (filenameOrError.code === "EACCES") filenameOrError.code = "EPERM";
-
-      this.emit(eventType, filenameOrError);
-    } else {
-      this.emit("change", eventType, filenameOrError);
-      this.#listener(eventType, filenameOrError);
-    }
-  }
-
-  close() {
-    this.#watcher?.close();
-    this.#watcher = null;
-  }
-
-  ref() {
-    this.#watcher?.ref();
-  }
-
-  unref() {
-    this.#watcher?.unref();
-  }
-
-  // https://github.com/nodejs/node/blob/9f51c55a47702dc6a0ca3569853dd7ba022bf7bb/lib/internal/fs/watchers.js#L259-L263
-  start() {}
-}
 
 function openAsBlob(path, options) {
   return Promise.$resolve(Bun.file(path, options));
@@ -577,7 +504,7 @@ var access = function access(path, mode, callback) {
   Dirent = fs.Dirent,
   Stats = fs.Stats,
   watch = function watch(path, options, listener) {
-    return new FSWatcher(path, options, listener);
+    return require("internal/fs/watch").watch(path, options, listener);
   },
   opendir = function opendir(path, options, callback) {
     // TODO: validatePath
@@ -1218,7 +1145,6 @@ function setName(fn, value) {
   Object.$defineProperty(fn, "name", { value, enumerable: false, configurable: true });
 }
 setName(Dirent, "Dirent");
-setName(FSWatcher, "FSWatcher");
 setName(Stats, "Stats");
 setName(_toUnixTimestamp, "_toUnixTimestamp");
 setName(access, "access");
